# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SILICOFELLER Quantum Studio API",
        "description": "AI-augmented quantum hardware EDA platform — V2. DesignGraph → Constraints → Placement → Routing → DRC → Qiskit Metal → Tapeout.",
        "version": "2.0.0"
    },
    "paths": {
...
        "/generate": {
            "post": {
                "tags": [
                    "generate"
                ],
                "summary": "Generate",
                "operationId": "generate_generate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Generate Generate Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/generate/frequency-plan": {
            "post": {
                "tags": [
                    "physics"
                ],
                "summary": "Frequency Plan",
                "description": "Standalone IBM-style frequency planning.\nReturns qubit A/B groups, EJ/EC, resonator band 6.3–7.2 GHz, λ/4 lengths,\ncollision warnings.",
                "operationId": "frequency_plan_api_generate_frequency_plan_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FrequencyPlanRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Frequency Plan Api Generate Frequency Plan Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/generate/drc": {
            "post": {
                "tags": [
                    "physics"
                ],
                "summary": "Drc Check",
                "description": "Run 7-rule DRC: spacing, CPW dimensions, frequency collision,\nresonator uniqueness, dispersive detuning, feedline clearance,\ngeometry overlap pre-check.",
                "operationId": "drc_check_api_generate_drc_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__routers__generate__DRCRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Drc Check Api Generate Drc Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/design/generate": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Generate Design",
                "description": "Full V2 constraint-driven design pipeline.\n\nPipeline: Constraints → Graph → Validation → Frequency Plan\n          → Placement → Routing → DRC → Qiskit Metal Code → Exports",
                "operationId": "generate_design_api_design_generate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConstraintsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Generate Design Api Design Generate Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/design/generate-from-graph": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Generate From Graph",
                "description": "Run the V2 pipeline from an existing graph JSON\n(e.g. drawn in the schematic editor).",
                "operationId": "generate_from_graph_api_design_generate_from_graph_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GraphDesignRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Generate From Graph Api Design Generate From Graph Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/design/from-prompt": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "From Prompt",
                "description": "NL prompt → DesignDocument via the Claude constrained-JSON + fallback ladder.\n\nUses the three-tier NL→graph pipeline:\n  1. Claude tool-call (forced DesignIntent JSON schema)\n  2. ML intent resolver fallback\n  3. Regex parse_prompt fallback\n\nReturns the full pipeline result including ``design`` (DesignDocument),\n``graph``, ``freq_plan``, ``drc``, ``code``, ``qclang``, and ``issues``\n(structured warnings from intent resolution and compilation).",
                "operationId": "from_prompt_api_design_from_prompt_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FromPromptRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response From Prompt Api Design From Prompt Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/design/validate": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Validate Graph",
                "description": "Structural graph validation — fast, no physics computation.",
                "operationId": "validate_graph_api_design_validate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GraphDesignRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Validate Graph Api Design Validate Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/design/route": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Route Placement",
                "description": "Run routing on a placement dict (from /api/generate/placement).\nReturns coupler, resonator, and feedline route segments.",
                "operationId": "route_placement_api_design_route_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RouteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Route Placement Api Design Route Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/design/drc": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Run Advanced Drc",
                "description": "Run the advanced 4-domain DRC (geometry, frequency, fabrication, connectivity).",
                "operationId": "run_advanced_drc_api_design_drc_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__routers__design__DRCRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Run Advanced Drc Api Design Drc Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GenerateRequest": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 2,
                        "title": "Prompt"
                    },
                    "substrate": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Substrate"
                    },
                    "metal": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metal"
                    }
                },
                "type": "object",
                "required": [
                    "prompt"
                ],
                "title": "GenerateRequest"
            },
...
            "GraphDesignRequest": {
                "properties": {
                    "graph": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Graph"
                    },
                    "constraints": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Constraints"
                    }
                },
                "type": "object",
                "required": [
                    "graph"
                ],
                "title": "GraphDesignRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/token"
                    }
                }
            }
        }
    }
}